libs/mysql-lib/pyproject.toml
[project]
name = "mysql-lib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "HoangManhh", email = "nhmanh4work@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "mysql-connector-python>=8.0.0",
]

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
